--- a/app/models/post_mover.rb
+++ b/app/models/post_mover.rb
@@ -418,8 +418,8 @@ class PostMover
   def update_statistics
     destination_topic.update_statistics
     original_topic.update_statistics
-    TopicUser.update_post_action_cache(topic_id: original_topic.id)
-    TopicUser.update_post_action_cache(topic_id: destination_topic.id)
+    TopicUser.update_post_action_cache(topic_id: original_topic.id, post_id: @post_ids, following_move: true)
+    TopicUser.update_post_action_cache(topic_id: destination_topic.id, post_id: @post_ids, following_move: true)
   end
 
   def update_user_actions