@ZogStriP 2021-07-12
> On an example site, this reduced the query **from ~6s to <10ms**
> [...]
> On an example site this reduced the `update_post_action_cache` time **from ~27s to 300ms**

👏